pyproject.toml
[project]
name = "tl-dm"
version = "0.1.0"
description = "Too Long; Did Meet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "litellm==1.82.4",
    "google-api-python-client==2.193.0",
    "ffmpeg-python==0.2.0",
]

[project.scripts]
tldm = "tldm.cli:main"

[dependency-groups]
dev = [
    "ruff>=0.14,<0.15",
    "pytest>=9.0,<9.1",
    "pytest-asyncio>=1.3,<1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["tldm"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "BLE",      # flake8-blind-except
    "D",        # pydocstyle
    "EM",       # flake8-errmsg
    "FBT",      # flake8-boolean-trap
    "FIX",      # flake8-fixme
    "TC",       # flake8-type-checking
    "TD",       # flake8-todos
    "TRY",      # tryceratops
    "SIM108",   # if-else-block-instead-of-if-exp
    "COM812",   # missing-trailing-comma (due to conflicts with formatter)
    "E501",     # line-too-long
    "T201",     # print-statement
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.pylint]
max-args = 15

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = [
    "ANN",      # flake8-annotations
    "S",        # flake8-bandit
    "SLF",      # flake8-self
    "ARG001",
]
"**/config.py" = [
    "S105",     # Possible hardcoded password (dev defaults)
]
